# Improve performance of bad_times and mask_times